refactor: report into a module
