+++ oceanarray/report/__init__.py
@@ -1,0 +1,5 @@
+"""Mooring report package — public API."""
+
+from ._mooring import MooringReport
+
+__all__ = ["MooringReport"]
